requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.104.0
FastAPI-SQLAlchemy==0.2.1
gevent==23.9.1
greenlet==3.0.0
h11==0.14.0
httptools==0.6.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
Mako==1.2.4
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
protobuf==4.21.12
pydantic==2.4.2
pydantic_core==2.10.1
PyMySQL==1.1.0
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.21.0
websocket==0.2.1
websockets==12.0
zipp==3.17.0
zope.event==5.0
zope.interface==6.1